Moved keyboard file location
